+++ app/src/main/java/com/example/softecadminapp/DATA/StallsData.kt
@@ -2,9 +2,9 @@ package com.example.softecadminapp.DATA
 
 data class StallsData(
     var bidTitle: String,
-    var bidders: Int,
+    var bidders: Double,
     var description_1: String,
     var description_4: String,
-    var maxBid: Int,
-    var minBid :Int
+    var maxBid: Double,
+    var minBid :Double
 )
